Move The Logging Event Layout project from Platform services  to Daikon
