+++ daikon-logging/logging-event-layout/src/main/java/org/talend/daikon/logging/event/field/HostData.java
