+++ daikon-logging/logging-event-layout/src/main/java/org/talend/daikon/logging/event/layout/Log4j2JSONLayout.java
@@ -1,5 +1,5 @@
 
-package org.talend.daikon.logging.event;
+package org.talend.daikon.logging.event.layout;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
