+++ daikon-logging/logging-event-layout/src/test/java/org/talend/daikon/logging/event/LogBackJSONLayoutTest.java
@@ -6,6 +6,7 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.MDC;
+import org.talend.daikon.logging.event.layout.Log4jJSONLayout;
 
 import net.minidev.json.JSONObject;
 import net.minidev.json.JSONValue;
