+++ daikon-logging/logging-event-layout/src/test/resources/log4j2.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<configuration status="DEBUG" packages="org.talend.daikon.logging.event" verbose="false">
+<configuration status="DEBUG" packages="org.talend.daikon.logging.event.layout" verbose="false">
   <appenders>
     <Console name="Console" target="SYSTEM_OUT">
       <PatternLayout pattern="CONSOLE APPENDER %d{HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n"/>
